change return type of lintso to int form long

+++ zio-maelstrom/src/test/scala/com/bilal-fazlani/zio-maelstrom/testkit/LinTsoFake.scala
@@ -4,12 +4,12 @@ package testkit
 import zio.*
 import protocol.*
 
-case class LinTsoFake(ref: Ref[Long]) extends LinTso:
-  override def ts(messageId: MessageId, timeout: Duration): ZIO[Any, AskError, Long] =
+case class LinTsoFake(ref: Ref[Int]) extends LinTso:
+  override def ts(messageId: MessageId, timeout: Duration): ZIO[Any, AskError, Int] =
     ref.updateAndGet(_ + 1)
 
 object LinTsoFake:
-  val make = ZLayer.fromZIO(Ref.make(0L).map { r =>
+  val make = ZLayer.fromZIO(Ref.make(0).map { r =>
     val impl = LinTsoFake(r)
     new LinTso:
       export impl.*
